Rename adf_dev_ramdisk to adf_dev_driver_ramdisk, lclevy#73
