+++ src/CMakeLists.txt
@@ -53,6 +53,8 @@ add_library ( adf
   adf_dev_driver.h
   adf_dev_driver_dump.c
   adf_dev_driver_dump.h
+  adf_dev_driver_ramdisk.c
+  adf_dev_driver_ramdisk.h
   adf_dev_drivers.c
   adf_dev_drivers.h
   adf_dev_flop.c
@@ -60,8 +62,6 @@ add_library ( adf
   adf_dev.h
   adf_dev_hd.c
   adf_dev_hd.h
-  adf_dev_ramdisk.c
-  adf_dev_ramdisk.h
   adf_dir.c
   adf_dir.h
   adf_env.c
@@ -96,7 +96,7 @@ add_library ( adf
 
 set_target_properties ( adf PROPERTIES
     #PUBLIC_HEADER "adflib.h"
-    PUBLIC_HEADER "adf_bitm.h;adf_blk.h;adf_cache.h;adf_dev_driver_dump.h;adf_dev_flop.h;adf_dev.h;adf_dev_hd.h;adf_dev_ramdisk.h;adf_dir.h;adf_env.h;adf_err.h;adf_file_block.h;adf_file.h;adf_file_util.h;adflib.h;adf_link.h;adf_nativ.h;adf_raw.h;adf_salv.h;adf_str.h;adf_types.h;adf_util.h;adf_version.h;adf_vol.h;debug_util.h;defendian.h;hd_blk.h;prefix.h"
+    PUBLIC_HEADER "adf_bitm.h;adf_blk.h;adf_cache.h;adf_dev_driver_dump.h;adf_dev_driver_ramdisk.h;adf_dev_flop.h;adf_dev.h;adf_dev_hd.h;adf_dir.h;adf_env.h;adf_err.h;adf_file_block.h;adf_file.h;adf_file_util.h;adflib.h;adf_link.h;adf_nativ.h;adf_raw.h;adf_salv.h;adf_str.h;adf_types.h;adf_util.h;adf_version.h;adf_vol.h;debug_util.h;defendian.h;hd_blk.h;prefix.h"
     #PRIVATE_HEADER ...
     VERSION ${CMAKE_PROJECT_VERSION}
 #    SOVERSION ${PROJECT_VERSION_MAJOR}
