+++ src/Makefile.am
@@ -16,10 +16,10 @@ libadf_la_SOURCES = \
     adf_cache.c \
     adf_dev.c \
     adf_dev_driver_dump.c \
+    adf_dev_driver_ramdisk.c \
     adf_dev_drivers.c \
     adf_dev_flop.c \
     adf_dev_hd.c \
-    adf_dev_ramdisk.c \
     adf_dir.c \
     adf_env.c \
     adf_file_block.c \
@@ -40,11 +40,11 @@ adfinc_HEADERS = \
     adf_cache.h \
     adf_dev_driver.h \
     adf_dev_driver_dump.h \
+    adf_dev_driver_ramdisk.h \
     adf_dev_drivers.h \
     adf_dev_flop.h \
     adf_dev.h \
     adf_dev_hd.h \
-    adf_dev_ramdisk.h \
     adf_dir.h \
     adf_env.h \
     adf_err.h \
